modules/Mfg_Vision_Image_Upload/requirements.txt
azure-iot-device~=2.7.0
azure-core
azure-storage-blob
azure-storage-file-datalake==12.5.0